Refactor catalog into separate components
