+++ src/components/Catalog/DeleteEventModal/DeleteEventModal.jsx
@@ -1,15 +1,15 @@
 import PropTypes from 'prop-types';
 import {
-    Button,
-    Modal,
-    ModalBody,
-    ModalContent,
-    ModalOverlay,
-    ModalHeader,
-    ModalCloseButton,
-    ModalFooter,
-  } from '@chakra-ui/react';
-  import { NPOBackend } from '../../utils/auth_utils';
+  Button,
+  Modal,
+  ModalBody,
+  ModalContent,
+  ModalOverlay,
+  ModalHeader,
+  ModalCloseButton,
+  ModalFooter,
+} from '@chakra-ui/react';
+import { NPOBackend } from '../../../utils/auth_utils';
 
 const DeleteEventModal = ({ isOpen, onClose, deleteItemId, setModified }) => {
   const handleConfirmDelete = async idToDelete => {
@@ -24,25 +24,25 @@ const DeleteEventModal = ({ isOpen, onClose, deleteItemId, setModified }) => {
 
   return (
     <Modal isOpen={isOpen} onClose={onClose}>
-        <ModalOverlay />
-        <ModalContent>
-          <ModalHeader>Confirm Delete</ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>Are you sure you want to delete this row?</ModalBody>
-          <ModalFooter>
-            <Button onClick={() => handleConfirmDelete(deleteItemId)}>Delete</Button>
-            <Button onClick={onClose}>Cancel</Button>
-          </ModalFooter>
-        </ModalContent>
+      <ModalOverlay />
+      <ModalContent>
+        <ModalHeader>Confirm Delete</ModalHeader>
+        <ModalCloseButton />
+        <ModalBody>Are you sure you want to delete this row?</ModalBody>
+        <ModalFooter>
+          <Button onClick={() => handleConfirmDelete(deleteItemId)}>Delete</Button>
+          <Button onClick={onClose}>Cancel</Button>
+        </ModalFooter>
+      </ModalContent>
     </Modal>
   );
-}
+};
 
 DeleteEventModal.propTypes = {
   isOpen: PropTypes.bool.isRequired,
   onClose: PropTypes.func.isRequired,
   deleteItemId: PropTypes.number.isRequired,
   setModified: PropTypes.func.isRequired,
-}
+};
 
 export default DeleteEventModal;
