+++ src/components/Catalog/SearchFilter/SearchFilter.jsx
@@ -1,0 +1,5 @@
+const SearchFilter = () => {
+  return <div>SearchFilter</div>;
+};
+
+export default SearchFilter;
