+++ src/pages/Catalog/Catalog.jsx
@@ -1,13 +1,5 @@
 import {
-  Table,
-  Thead,
-  Tbody,
-  Tr,
-  Th,
-  Td,
   Text,
-  TableContainer,
-  Badge,
   Input,
   Select,
   Flex,
@@ -16,14 +8,14 @@ import {
   useDisclosure,
   Container,
   Link,
+  TableContainer,
 } from '@chakra-ui/react';
 import { AddIcon, SearchIcon } from '@chakra-ui/icons';
-import s from './Catalog.module.css';
 import { useState } from 'react';
-import PaginationFooter from '../../components/PaginationFooter/PaginationFooter';
-import { EditIcon, DeleteIcon } from '@chakra-ui/icons';
-import DeleteEventModal from '../../components/DeleteEventModal/DeleteEventModal';
-import CreateEventFormModal from '../../components/CreateEventForm/CreateEventFormModal';
+import DeleteEventModal from '../../components/Catalog/DeleteEventModal/DeleteEventModal';
+import CreateEventFormModal from '../../components/Catalog/CreateEventForm/CreateEventFormModal';
+import PaginationFooter from '../../components/Catalog/PaginationFooter/PaginationFooter';
+import CatalogTable from '../../components/Catalog/CatalogTable';
 
 const subjectsOptions = [
   'life skills',
@@ -211,7 +203,6 @@ export default function Catalog() {
             </Link>
           </Flex>
         </Flex>
-
         <TableContainer
           mt="8"
           mb="8"
@@ -220,129 +211,11 @@ export default function Catalog() {
           borderWidth="1px"
           borderRadius="12px"
         >
-          <Table
-            variant="simple"
-            className={s['catalog-table']}
-            borderWidth="0"
-            width="100%"
-            px="1rem"
-          >
-            <Thead>
-              <Tr>
-                <Th textAlign="left">Event</Th>
-                <Th textAlign="left">Host</Th>
-                <Th textAlign="left">Tags</Th>
-                <Th width="8rem"></Th>
-              </Tr>
-            </Thead>
-            <Tbody>
-              {tableData.map(
-                ({ id, host, title, eventType, subject, year, season, description }) => (
-                  <Tr key={id}>
-                    <Td textAlign="left" py="1.5rem">
-                      {title}
-                    </Td>
-                    <Td textAlign="left" py="1.5rem">
-                      {host}
-                    </Td>
-                    <Td textAlign="left" py="1.5rem">
-                      <Container
-                        p="0"
-                        m="0"
-                        maxW="16rem"
-                        display="flex"
-                        flexWrap="wrap"
-                        gap="0.375rem"
-                      >
-                        {season && (
-                          <Badge
-                            backgroundColor="#CEECC3"
-                            color="gray.900"
-                            textTransform="capitalize"
-                            borderRadius="10rem"
-                            fontWeight="normal"
-                            px="0.5rem"
-                            mr="0.125rem"
-                          >
-                            {season}
-                          </Badge>
-                        )}
-                        {year && (
-                          <Badge
-                            backgroundColor="#FFE1BE"
-                            color="gray.900"
-                            textTransform="capitalize"
-                            borderRadius="10rem"
-                            fontWeight="normal"
-                            px="0.5rem"
-                            mr="0.125rem"
-                          >
-                            {year}
-                          </Badge>
-                        )}
-                        {subject && (
-                          <Badge
-                            backgroundColor="#E8D7FF"
-                            color="gray.900"
-                            textTransform="capitalize"
-                            borderRadius="10rem"
-                            fontWeight="normal"
-                            px="0.5rem"
-                            mr="0.125rem"
-                          >
-                            {subject}
-                          </Badge>
-                        )}
-                        {eventType && (
-                          <Badge
-                            backgroundColor="#CFDCFF"
-                            color="gray.900"
-                            textTransform="capitalize"
-                            borderRadius="10rem"
-                            fontWeight="normal"
-                            px="0.5rem"
-                            mr="0.125rem"
-                          >
-                            {eventType}
-                          </Badge>
-                        )}
-                      </Container>
-                    </Td>
-                    <Td>
-                      <IconButton
-                        color="gray.400"
-                        backgroundColor="transparent"
-                        p="0.5rem"
-                        h="fit-content"
-                        w="fit-content"
-                        icon={<EditIcon />}
-                        onClick={() =>
-                          handleEditForm({
-                            id,
-                            title,
-                            host,
-                            year,
-                            eventType,
-                            subject,
-                            description,
-                          })
-                        }
-                      />
-                      <IconButton
-                        color="gray.400"
-                        backgroundColor="transparent"
-                        p="0.5rem"
-                        h="fit-content"
-                        w="fit-content"
-                        icon={<DeleteIcon />}
-                        onClick={() => handleDeleteClick(id)}
-                      />
-                    </Td>
-                  </Tr>
-                ),
-              )}
-            </Tbody>
-          </Table>
+          <CatalogTable
+            tableData={tableData}
+            handleEditForm={handleEditForm}
+            handleDeleteClick={handleDeleteClick}
+          />
           <PaginationFooter
             table="catalog"
             setData={setTableData}
@@ -351,6 +224,7 @@ export default function Catalog() {
             isModified={isModified}
           />
         </TableContainer>
+
         <CreateEventFormModal
           isOpen={isCreateFormOpen}
           onClose={onCreateFormClose}
